Implemented sequentially thresholded least squares. Made this the default optimization algorithm for SINDy. Moved this and other optimization functions (including STRidge) to new files called optimizers.py
